司机接单

+++ src/main/java/cn/deliver/dao/UserDao.java
@@ -16,7 +16,7 @@ public interface UserDao {
 
     int insertSelective(User record);
 
-    User selectByPrimaryKey(Long id);
+    User selectByPrimaryKey(Integer id);
 
     int updateByPrimaryKeySelective(User record);
